Move Hall Management to admin
