+++ src/Pages/Admin/Dashboard/Dashboard.jsx
@@ -65,7 +65,7 @@ function Dashboard() {
           <Link type="button" className="btn btn btn-outline-success" to="/admin/dashboard/pending-requests"> Pending Reservations</Link>
         </div>
         <div className="col-md-3 mt-1 col-12 d-flex justify-content-center align-content-center">
-          <Link type="button" className="btn btn btn-outline-success" to="/admin/halls"> Manage Halls</Link>
+          <Link type="button" className="btn btn btn-outline-success" to="/admin/dashboard/halls"> Manage Halls</Link>
         </div>
 
       </div>
