+++ src/Pages/Admin/Hall/components/create.jsx
@@ -1,8 +1,8 @@
 import Swal from 'sweetalert2';
 import { useState } from 'react';
 import { FaPlus } from 'react-icons/fa';
-import { addHall } from '../../../Api/admins/halls/api';
-import Modal from '../../../components/Modal';
+import { addHall } from '../../../../Api/admins/halls/api';
+import Modal from '../../../../components/Modal';
 import '../Hall.css';
 
 const Create = ({ setChanged }) => {
